Make sure clearing the context doesn't clear previous_context
Creating a shallow copy left the previous_context's  __dict__ as the same object as the current context's __dict__, so clearing the current context also cleared the previous_context. To solve it, we just creat a copy of the __dict__ itself, and then use that copy to restore the previous context.