Create tasks for AWS Lambda interaction
Tasks include

- Create
- Delete
- Invoke
- List

Each task serves as a wrapper around a boto3 Lambda client.
Defaults are set to make use as user friendly as possible.

Each task is designed to take advantage of the full feature
set of boto3, and all possible arguments are included.

Because in many cases there isn't a good 'null' argument
equivalent to pass into the boto3 client, logic is in place
to control which arguments get passed.

AWS credentials are passed to the boto3 cleint in the same
manner as existing S3 upload/donwload tasks.

Create

Task to create a Lambda function from a local zip file or
an S3 object. Logic is such that if a path is provided
to a zip file, S3 related arguments are disregarded.

Delete

Task to delete a Lambda function, specified by name
and optionally by version. In the run method, logic is
Task to delete a Lambda function, specified by name
and optionally by version. In the run method, logic is
used to ensure the Qualifier is only passed to boto3 if
specified in the constructor.

Invoke

Task to invoke a Lambda function, specified by name and
optionally by version. Other optional parameters can be
passed as documented. One point that may need clarity is
around *client_context*. As currently implemented, the user
can pass a Python dictionary, and it is transformed into
a base64 encoded json file. Other options are available,
like forcing the user to pass a JSON object, or checking
the input object type and adjusting as necessary.

The run method allows user to override object's payload
value.

List

Task to list all Lambda functions. Defaults are implemented
to list the broadest possible set of functions. As currently
implemented, the user will need to manually implement the
pagination process for extensive quantities of Lambda functions.

Resolves #1009. Accompanying test cases in next commit.