Add spaCy to task library
New directory and import logic for spaCy tasks. This
will be an extra component and we will raise an ImportError
if prefect is not installed with the spaCy extra.

SpacyNLP

This is a task for processing text with a spaCy language
model. Text can be provided at run, allowing users to
use the same model to create mulitple Documents.

The spaCy language model can be fed in two different ways.

With the 'nlp' parameter, you can specify an entire
custom pipeline to use. My assumption is most sophisticated
users of spaCy would go this route, and it's not worth building
in extensive logic for spaCy customization operations.

Alternatively, you can load a default spaCy model (if you have
downloaded it) with the spacy_model_name parameter. We also
give the ability to pass through arguments 'disable' and
'component_cfg' to spacy.load(). An error is raised if the
model fails to load.

SpacyTagger, SpacyParser, SpacyNER

As discussed with Josh, these tasks allow users to access
components of a spaCy pipeline. In default pipelines, these
are attributes that can be easily accessed.

The pipeline parameter, 'nlp', can be specified at run time.

But Zach, what if I want to access other pipeline components?

SpacyComponent (!!)

spaCy language pipelines can have all sorts of custom features.
These are stored in the 'pipeline' attribute as a list of tuples,
('feature_name', FeatureObject). This task allows users to
access custom component objects by name.

Both the name of the component and the pipeline can by specified
at run time.

Resolves issue #1018