Add tests for spaCy tasks
Includes basic testing for spaCy tasks.

1. Object construction
The construction of all objects is tested, including
a basic default value assertion.

2. Error handling (SpacyNLP)
The SpacyNLP class is also tested to ensure an error
is raised when you provide an un-loadable model name.

3. Accessing pipeline components
SpacyTagger, SpacyParser, SpacyNER, and SpacyComponent are
all tested using Mocks to ensure they are accessing the
correct properties when the task is run.

SpacyComponent uses the same structure as a spaCy pipeline,
[('component_name', Component)...]

Tests were run locally to ensure the default spaCy model
worked with SpacyNLP. I chose not to include spaCy models
in testing because of the additional dependencies required.
Resolves issue #1018