Add Postgres tasks to library
__init__.py follows standard tasklib style, raises exception on
ImportError informing the user that the 'postgres' extra must be
installed. The dependency is reflected in the extras section of
setup.py file.

Two tasks for interacting with postgres have been added. One to
execute a query and one to fetch results of a query.

ExecuteTask

User provides connection parameters needed for psycopg2.

Connection is established at run-time in order to properly manage
the connection object. It is important to call connection.close()
while we still have a reference to the connection.

The run method allows non-database specific parameters to be passed
at run time. It checks to ensure that a query string has in fact
been provided.

Any errors in establishing the connection are thrown without being
handled by prefect. There is a wide variety of such errors, and
generally I've found it is helpful to have the specific error from
psycopg2. I'm open to improvements here.

Once a connection is established, the run method uses context
management for the connection and the cursor objects. This ensures
that failed transactions are rolled back automatically, avoiding
future complications. Any exceptions during execution are passed
through, with the additional step of ensuring the connection is closed.

FetchTask

This task allows the user to get results from their query. All of
the above logic in ExecuteTask is also implemented here.

The addition includes two parameters that can be provided during
construction or run, fetch and fetch_count. These two parameters
control how results are fetched from the cursor object.

Checking is done to ensure a correct value has been passed into fetch.

Resolves #1022