Remove sentence saying CloudResultHandler is default
It looks like CloudResultHandler was removed as the default in the 0.6.0 release. (See https://github.com/PrefectHQ/prefect/releases/tag/0.6.0 and https://github.com/PrefectHQ/prefect/pull/1198)

As a healthcare company dealing with HIPAA/PHI data, it was scary to read this sentence (that I'm proposing to remove) in the docs: "As of this writing, the default handler that ships with a standard install of Prefect is the CloudResultHandler that writes data to a Prefect-managed Google Cloud Storage bucket." It caused me to worry that in testing Prefect Core we might have unwittingly shipped PHI data to a Google Cloud Storage bucket managed by Prefect.