Add optional HTTP health check to all agents [ci skip]
This adds an optional HTTP server to all agents (default is no server).
Currently this server only serves one route - `/api/health` - which can
be used by orchestration systems to do health checks. If the agent is
running it's assumed to be healthy, which should be a good default for
all systems (e.g. if the k8s api server goes down briefly for
maintenance, the agent pod shouldn't be marked as unhealthy because of
it). Those kind of health checks should be handled by communication with
cloud/server, and possibly reducing/stopping pulling new flows while the
backing infrastructure is unavailable.