Fix bug in local agent process cleanup
Previously this would loop over a list of processes, and *attempt* to
remove completed processes with `processes.pop`. Due to a bug though,
this would remove the incorrect process. We now use a set of processes,
which provides `O(1)` access to remove (more efficient than `O(n)`), and
removes the correct process.