Update changelog process
Instead of having every PR modify `CHANGELOG.md` themselves (which
easily led to merge conflicts), we now store changes in the `changes/`
directory. Each PR adds a single file to the directory describing the
change. On release, the `update_changelog.py` script can be run, which
compiles the changelog and updates `CHANGELOG.md` (and clears the
changes directory to restart this process). An `EXAMPLE.yaml` changelog
entry will persist between clears to provide some documentation of this
process.