[WIP] Reduce size of tasks submitted via prefect [ci skip]
Currently prefect submits very large objects to dask for every call.
Large objects are expensive to serialize and send between workers, which
results in slower execution and increased memory usage. The
depth-first-execution PR helped some, but there's still more we can do
to reduce this.

This PR changes the `run_task` method to be a function, and minimizes
the set of arguments submitted to that function. Things work fine in my
brief personal tests, but tests are failing (several of them due to the
tests monkeypatching internals that have changed, which is brittle).
Still need to update the tests, and fix any bugs.