Use pagination in k8s resource manager
Previously we'd do a full list in each loop of `clean_resources`. Under
heavy load, this may return a large list, which can use more resources
than desired. We now use the paginated api, listing in sets of 20.

We also explicitly set `propagationPolicy="Foreground"` when deleting
jobs to ensure that child objects are also deleted and not orphaned.
Afaik this shouldn't be necessary, but for some reason is on my local
kubernetes install (1.15.5). No harm is done by setting it though.