Add `executor` kwarg to environments
This allows configuring an instantiated `Executor` on an `Environment`.
With the addition of support for configuring Dask clusters inside the
`Executor`, this should remove the need for any dask-specific
environments. After deprecation, only the following environments should
remain:

- `LocalEnvironment`
- `KubernetesJobEnvironment`
- `FargateTaskEnvironment`

Here's an example of running a flow, where the flow-runner is a local
process, using a temporary dask cluster running on fargate:

```python
flow.environment = LocalEnvironment(
    executor=DaskExecutor(
        cluster_class="dask_cloudprovider.FargateCluster",
        cluster_kwargs={...},
    )
)
```

One nice thing about using an *instantiated* executor is that execution
locally with `flow.run()` and registering a flow to be executed later
with `flow.register()` both can use the same object. Since an executor
doesn't use any resources until its `.start()` method is called, there
should be no harm done using an instantiated executor in the
`Environment` object.