Reconfigure `run_flow`
Previously `run_flow` was a method on the class, which led to numerous
bugs. This moves that logic to a function `load_and_run_flow()`, which
loads the flow, then calls `flow.environment.run(flow)`. This simplifies
the running logic, and ensures the proper configuration is used. Tests
have been updated to follow.