Executors stop/wait for background work on exit.
Previously, the dask executors could leave work continuuing in the
background upon exit from the `start` contextmanager. This could lead to
odd behavior, where tasks from a flow run might still be running after
`flow.run()` has completed.

This PR changes the semantics of the `start` contextmanager so that no
lingering work remains after an executor has exited the `start`
contextmanager.

- For a `DaskExecutor` with a temporary cluster, we shutdown the
  cluster, so no lingering work can persist.
- For a `DaskExecutor` with an external cluster (specified with an
  `address`), we stop all pending tasks and wait for all running tasks
  to finish.
- For a `LocalDaskExecutor`, we terminate the backing pool as quickly as
  possible, then wait for the pool to close.
  - For a process pool, this terminates the processes
  - For a thread pool, we attempt to interrupt the threads (CPython
    specific).
- A `LocalExecutor` runs everything in the main thread, so no background
  tasks can persist.

This is a precursor to getting Cancellation to work in a robust way.