Same Python exe for is_serializable subprocess
This fixes an issue where if the user is running a Python that is not
the default Python on their $PATH then is_serializable fails because
it invokes a subprocess running a different Python (the system
default). This would occur if e.g. the user is running the executable
from a virtual env without first activating the virtual environment.