Improve output format in prefect describe cli
Previously the default format for `prefect describe` was to print the
python object. This is similar to json, but not actually json. There's
not a good reason to not use json, so we default to json formatting, and
also add yaml as an option. We also pretty-print the json to make it
easier to read.