Add Resume Failing Flow to Advanced Tutorials
My Stackoverflow question became far too long very
quickly.  I felt that condensing it into a more
useful markdown file would be of better service
to others.

Please feel free to ignore this if covered better
elsewhere in the docs, or if you think this is
worthwhile please get in touch! :)