Ensure full context is available in cancel-checking thread
Since `prefect.context` is a thread local, we need to ensure it's
properly copied into the background cancellation checking thread, since
logging relies on some parameters in the context. This wasn't caught
before since logging wasn't enabled. We fix this to copy the context
over to the background thread, and add a test to check that logging
calls all have the context available.