Optimizing constants out of graph still add downstream
Although an edge can be omitted from the graph if source task is a constant, the target task should still be included in the graph. As it is a single edge graph from constant to task will contain no tasks if "add_edge()" is used to build it, but it should contain at least the target task.