KubernetesAgent retries on connection errors
For errors with the connection (connection dropped, timeouts, etc...) we
can retry submitting jobs a few times. If a previous submission
succeeded (the job was created) but then errored to the client
additional submission attempts will return a 409 error, which we now
interpret as a success. This should help when deployed in setups with
network issues.

For now we retry a max of 3 times, with a fixed 1 second delay between
each attempt.