Fixup docs tooling/test
The docs test for properly formatted section headers was too strict, and
prevented use of header words anywhere else in the docstring (i.e.
"Example" used in a sentence was forbidden). We now only check if a
header keyword is on a line by itself without a ":". This should catch
the most common source of errors, forgetting the colon, without causing
too much trouble elsewhere.