Don't set log level in k8s agent w/ KubernetesRun
When using the new `KubernetesRun` config, we don't set the logging
level by default in the k8s agent. This makes it easier for users to
configure the logging level, as the agent doesn't always override it.
We only apply this change to flows using the `KubernetesRun` run
conifiguration as it's new and experimental. Users using other config
mechansisms may be relying on the existing behavior.