Better error for `flow.register` in flow context
Previously this would raise an error due to pickle failing. We could
have fixed it so pickle worked, but I'd argue that the flow context
should always exit before registering the flow (we do correctness checks
for the user on flow context exit). We now raise a nicer error message.