[client] Create a property active_tenant_id to wrap the tenant initialization.
Developpers using prefect server needs sometimes to attach_headers to the
queries in order to communicate with the server (e.g Basic Auth).

However, the first call to the server happened inside the constructor with the
tenant_id initialization. Since the headers couldn't been attached it results to failures.
This commit fix this issue.

The active_tenant_id property is introduced. Whenever needed the active_tenant_id
will be called. If not set the initialization will be perform.

close #3515