Nice warning if passing task to Task constructor
A common user mistake when usint class-based tasks is to pass dependent
tasks as arguments to the task constructor, rather than when calling the
task (or using methods like `.map`/`.bind`/`.set_dependencies`). We now
raise a nice warning in this case. We don't error outright in case the
user actually intended this (although my guess is this case is rare).

Also cleans up a few linting errors in the `test_task.py` file.