Add a new postgres task PostgresExecuteMany
PostgresExecuteMany is a Task subclass that use the psycopg2 driver to
make many queries against a Postgres database.

PostgresExecuteMany is an answer for the use case of looping over a
list of tuples.

The traditional method of tackling this use case was by wrapping
a loop over PostgresExecute which tended to:
	1- be slow
	2- fill the stdout with warnings