Hide private methods in api docs
Previously any non `_`-prefixed method in a class was documented as part
of the api docs. For things like agents (which are more an application
than a class), there were several non-public methods that aren't
`_`-prefixed. Having these in the api docs led some users to be
confused, thinking they could call those internal methods directly.

We now add a way to specify a set of methods to document, and apply this
to the agent api docs to only document the `start` methods.