Support flow's with null `run_config`
This adds support for flows with no `environment` AND no `run_config`.

Currently all flow's have an environment created by default. If a user
explicitly sets a `run_config` on a flow, the new `run_config` based
configuration is used, otherwise the default `environment` is used.

In 0.14.0 we plan to remove the default `environment`, so flow's will by
default have `run_config=None, environment=None`. This change adds
support for that, without removing the default (yet).

The new logic is now:

- If a flow has a run-config configured, use that. Agents check that the
  run-config set on the flow matches the type they support.
- If a flow has no environment, the agent will create the default
  `RunConfig` for that agent type. This means that a flow registered
  with no `run_config` configured on it is runable by all agents
  (storage and labels dependent), matching the existing behavior.
- If a flow has an `environment`, the old environment-based code path is
  taken.