Fix cycle creation in apply_map
Fixes accidental cycle creation in `apply_map` in cases where a `case`
statement is mapped over a task that isn't used elsewhere in the
sub-flow beforehand. This was due to the args to `apply_map` not being
added to the subflow before the mapped function is called, causing the
logic in the `case` statement to automatically add erroneous edges to
these tasks.