Avoid race condition when creation task run artifacts
Currently when running with Prefect Cloud, there's a rare race condition
that may occur when creating a task run artifact against a mapped task,
leading to errors. For now, we catch this error and retry a few times
with a fixed delay in between attempts.

No test added (hard to test, temporary fix), but I've confirmed this fix
works locally.