Use custom `NotImplementedError` in results
The `TaskRunner` does different things dependent on if a task's result
type has a `write` method implemented. Previously this would check for
`NotImplementedError`, which would shadow any actual
`NotImplementedError` raised when writing a result (e.g. in the
serializer). We now use a custom `ResultNotImplementedError` to handle
this control flow, avoiding the issue.