Better support multiple flows per file
We previously added support for storing multiple flows in a single
script file, but in practice that would fail since which flow was
retrieved was non-deterministic (the first flow found would be used).
This fixes this to always load and return the correct flow.

To accomplish this (and clean up some storage internals), the signature
for the `get_flow` method has been changed to accept the *flow name*
rather than the *flow location*. A location may host more than one flow,
but a flow name is always unique within a `Storage` object. Since the
`get_flow` method isn't user facing this change should be fine.