doc update: instantiate before passing PrefectSecret
The existing code examples do not work; they trigger the message:

UserWarning: A Task was passed as an argument to DatabricksSubmitRun, you likely want to first initialize DatabricksSubmitRun with any static (non-Task) arguments, then call the initialized task with any dynamic (Task) arguments instead.

Since PrefectSecret is a task, it cannot be passed when instantiating.