Support `EXTRA_PIP_PACKAGES` in docker image
This adds support for a `EXTRA_PIP_PACKAGES` environment variable in our
official `prefect` docker images. This takes a space delimited list of
packages to pass to `pip install` on image startup, allowing for easier
runtime installation of extra packages. For production deployments users
should still build images with their required dependencies, but for
development this can make things a bit easier.

Also extends our default entrypoint to drop the user into a shell if no
CMD is specified (with a header explaining what the image is). We've had
a few users confused with what the prefect image was for (mostly
assuming it would start prefect server automatically), so hopefully this
helps remove this confusion.