Ensure agent health-check thread has an active event loop
This was broken in 0.14.7. Previously we had accidentally relied
distributed implicitly setting a global `AnyThreadEventLoop` policy on
import, but in 0.14.7 we dropped all unnecessary imports (so this
implicit global change didn't happen). Our tests still passed though,
since `distributed` is imported as part of the test suite.

This bug caused all health checks to fail in the agent, since the event
loop wouldn't properly start. I don't have a good test for this (global
state makes testing tricky), but have verified locally that this fixes
the issue.