Adds FivetranSyncTask to manage Fivetran connectors via Prefect
The FivetranSyncTask allows a Fivetran user to manage a connector
sync process through Prefect. This requires a Fivetran account
(https://fivetran.com) and  a connector configured through the
Fivetran user interface, (https://fivetran.com/dashboard).

This Task will attempt to raise an error on any bad values, HTTP
status codes, and on a failed sync process. Otherwise, it will
end successfully and returned the timestamp that the sync ended
successfully.

This is a very early-stage attempt at this Task, feedback is very
welcome!