Make slug generation more robust
Previously generating a slug for a new task was an O(n) operation where
all existing tasks were inspected to determine the next slug to use for
the new task. Besides being mildly expensive, this wasn't robust to
tasks changing name after being added to a flow, leading to duplicate
slugs. We now track auto-generated slugs in a private member on the flow
class, and explicitly handle clashes.

Note that this is still technically O(n) (since we check for a clash in
`slugs.values()`, which is a linear search). But the operation takes
place entirely within the Python C-api, so is generally fast for the
size of `n` we typically deal with.