Don't forward nout in mapped tasks
Previously setting `nout` on a task would forward the `nout` to both
normal task call outputs (which we want) and mapped task outputs (which
we don't). This led some users to assume that calling `task.map` would
result in a tuple of lists of results, when the actual output is a list
of tuples.

This PR fixes this to stop propagating the `nout` field on `task.map`
calls, so the improper usage will error instead of implicitly passing.