Register a new task definition for every ECS execution
Due to our new support for per-run run configs, we no longer have a way
to cache duplicate registered task definitions for a flow version (only
registering a new ECS task definition if the flow version updates).
Since ECS task definitions don't cost anything, and are quick to
create, we switch to creating a new definition for every flow run. The
definition is immediately deregistered after the task is created, so the
cluster shouldn't become cluttered with 1000s of unused task
definitions.