Support setting `image` in agent-side templates
Previously we always overrode the `image` used at runtime for the K8s
and ECS agents, which prevented setting a default image to use on the
agent. This PR fixes that, allowing agent side defaults to be set
(through `--job-template` on the k8s agent and `--task-definition` on
the ECS agent).