Add ability to stream ShellTask logs with level INFO
There are certain circumstances under which we'd like to be able to stream
the output of a `ShellTask` (or its child classes) without having to set the
global logging level to `DEBUG`; for example, the process launched by `ShellTask`
sometimes streams valuable output that a user might like to capture and stream
using the base Prefect logger. Currently, this is not possible. This change
gives the user more control in that they can change the log level from one
`ShellTask` to the next within a given flow.