Change ShellTask.stream_output to type Union[bool, int]
Rather than having two kwargs to handle if and how to stream ShellTask logs,
consolidate these into one argument (the originally named `stream_output`) so
that an integer can be passed as an argument as well as a boolean. The integer
argument allows the user to set the level they'd like Prefect to use
when logging the `ShellTask`'s output stream.