Do not set labels as an empty set during build/register
An empty set has a different meaning than `None` which means that
flows built without labels in this setup cannot be run on an agent
unless it also has explicitly empty labels. It's better to have
these flows be runnable on any agent if labels are not provided by
the user.

Note that the `labels.union` change is merely for a shorter line
by avoiding an unecessary `set` cast since `.union` takes any
iterable as an arg