A few register CLI fixes (#4593)
* Use absolute path when loading scripts in register

Previously this would erroneously use a relative path when a relative
path was provided, which required the agent to be run in the same
directory as the register command was run in. We now normalize all paths
to absolute.

* Add storage labels to flow in register CLI

Previously these labels were accidentally omitted. This only affects
flows registered using `Local` storage, since that's the only storage
that defines labels by default.

* Add changelog entry

* Fix windows tests