Use server_onupdate=FetchedValue()
We use a server-generated timestamp for the "updated" column,
so we need to mark the column as `server_onupdate=FetchedValue()`
to ensure that SQLAlchemy includes this column in the `RETURNING`
clause for databases that support this feature.

Without the column in a `RETURNING` clause, SQLAlchemy needs to
immediately `SELECT` the row again to get the generated updated
time, leading to unnecessary query activity.

Fixes PrefectHQ/orion#674.