Allow postgres tasks to configure all params at runtime (#5016)
* Allow postgres tasks to configure all params at runtime

This change allows the Postgres tasks to be fully configured at runtime as well as init. 

I followed the lead of: https://github.com/PrefectHQ/prefect/pull/4545/ and https://github.com/PrefectHQ/prefect/pull/4907

* Fix copy/paste issue where I changed list to tuple

I had a copy/paste issue in my initial commit that changed the `data` argument in PostgresExecuteMany from a `list` to a `tuple`. Reverted the change to that line.

* Create pr5016.yml

* Format with black

* Fixup changes entry

Co-authored-by: Michael Adkins <michael@prefect.io>