Add the ability to plot graphs horizontally. (#5472)
* core: add horizontal parameter to visualize function

This commit introduces the option to plot or visualize a task graph
horizontally. If you plan to plot this figure in a doc for instance, it
can be large if you plot vertically. It is better to have the option to
plot in both orientations.

Signed-off-by: Julio Faracco <jcfaracco@gmail.com>

* tests: add test to check if graphviz plot horizontally

This is a simple coherence check to verify the new visualize parameter
works fine.

Signed-off-by: Julio Faracco <jcfaracco@gmail.com>

* changes: add changes file

Signed-off-by: Julio Faracco <jcfaracco@gmail.com>

* Run `black`

Co-authored-by: Michael Adkins <michael@prefect.io>