Use pending state when flow run state is missing (#5554)
* Use pending state when flow run state is missing

There is a slight timeframe in which an initialized flow run
does not have a serialized state. During this time a backend
FlowRunView object cannot be initiated, because None values
for serialized_state raises exception. This change handles
null values, and fills with  a Pending state with an informative
message.

* Add change entry for PR #5554