Fix bug where futures would be marked incorrectly marked as async
Task run futures track if they are asynchronous in order to
determine if `result()` should return a coroutine or retrieve the
result with a blocking call. When we added the ability to call
async tasks from sync flows, we did not update this tracking to
take this into account. Accordingly, if you call future.result()
for an async task in a sync flow you will receive a coroutine
instead of the result of the task run.