Add inference of Docker network mode for 'host' and 'none' networks
This brings our behavior in-line with the Docker CLI which will
change the network mode on created containers to 'host' or 'none'
respectively if the special network names 'host' or 'none' are used.

These networks appear to be immutably defined by Docker.

'host' networking is still not expected to work on non-Linux systems,
see the Docker documentation for more information.

Closes https://github.com/PrefectHQ/prefect/issues/4840
Closes https://github.com/PrefectHQ/prefect/issues/2638