Add No Ui Flag to Kubernetes Dev
When creating a kuberenetes manifest, the previous template did not include the no-ui option. This resulted
in the container crashing on load, as it does not have npm, which gets run as part of the UI standup process.

To overcome this, we launch orion without the UI in kube dev